backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
httpx==0.28.1
pydantic==2.10.5
pydantic-settings==2.7.1
python-multipart==0.0.20
chromadb==0.6.3
pypdf==5.1.0
python-docx==1.1.2
chardet==5.2.0
aiosqlite==0.20.0
